test_: move wait_for_login to StatusBackend
